[api/db] move wrapper package
